+++ CHANGELOG.md
@@ -43,11 +43,13 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Fix typeos in marker log statements [#502](https://github.com/provenance-io/provenance/issues/502)
 * Set default coin type to network default [#534](https://github.com/provenance-io/provenance/issues/534)
 * Add logger to upgrade handler [#507](https://github.com/provenance-io/provenance/issues/507)
+* Allow markers to be created over existing accounts if they are not a marker and have a zero sequence [#520](https://github.com/provenance-io/provenance/issues/520)
 * Removed extraneous Metadata index deletes/rewrites [#543](https://github.com/provenance-io/provenance/issues/543)
 
 ### Bug Fixes
 
 * When deleting a scope, require the same permissions as when updating it [#473](https://github.com/provenance-io/provenance/issues/473)
+* Allow manager to adjust grants on finalized markers [#545](https://github.com/provenance-io/provenance/issues/545)
 
 ## [v1.7.5](https://github.com/provenance-io/provenance/releases/tag/v1.7.5) - 2021-10-22